+++ ServerSide/CongestionControl/CongestionHandler.h
@@ -12,7 +12,7 @@
 
 class CongestionHandler {
 private:
-  int window_size;
+    float window_size;
   Slowstart * slowstart;
   CongestionControl* congestionControl;
   int threshold;
@@ -21,8 +21,10 @@ private:
 public:
     CongestionHandler();
     STATE getCurrentState();
-    int update_window_size(std::string event);
-    int get_curr_window_size();
+
+    float update_window_size(std::string event);
+
+    float get_curr_window_size();
 };
