+++ ServerSide/CongestionControl/Slowstart.h
@@ -10,17 +10,17 @@
 
 class Slowstart  {
 private:
-    int window_size;
+    float window_size;
     int threshold;
     STATE self;
 
 public:
-    Slowstart(int wsz, int threshold);
+    Slowstart(float wsz, int threshold);
     STATE update_window_size(std::string event);
     void set_Wsz(int window_size);
     void set_threshold(int threhold);
 
-    int getWindowSize();
+    float getWindowSize();
     int getThreshold();
